refactor(db): store money as DECIMAL(14,2) instead of FLOAT
chk_amount and the two account balance columns were FLOAT, which cannot
represent most cent values exactly: 12143 of 17496 transaction amounts
had representation error, and balances had drifted by $0.33 on one
account and $0.02 on another.

update_balances() compounded it by casting MySQL's exact SUM() to a PHP
float before writing it back. It now binds the decimal string through
unchanged; migrating the columns without this would have left the drift
accumulating.

+++ tables-mysql.sql
@@ -23,9 +23,9 @@ CREATE TABLE chk_account (
   /* account number */
   chk_account_no VARCHAR(25) NULL,
   /* Current balance */
-  chk_balance FLOAT NOT NULL DEFAULT 0,
+  chk_balance DECIMAL(14,2) NOT NULL DEFAULT 0,
   /* Current bank balance */
-  chk_bank_balance FLOAT NOT NULL DEFAULT 0,
+  chk_bank_balance DECIMAL(14,2) NOT NULL DEFAULT 0,
   /* Is account archived? (Y/N) */
   chk_is_archived CHAR(1) DEFAULT 'N',
   PRIMARY KEY ( chk_acct_id )
@@ -51,7 +51,7 @@ CREATE TABLE chk_trans (
   /* Check number */
   chk_no INT NULL,
   /* Amount (positive for deposit, negative for all else) */
-  chk_amount FLOAT NOT NULL,
+  chk_amount DECIMAL(14,2) NOT NULL,
   /* Date of transaction (YYYYMMDD) format */
   chk_date INT NOT NULL,
   /* Description */
@@ -93,7 +93,7 @@ CREATE TABLE chk_bank_trans (
   /* Check number */
   chk_no INT NULL,
   /* Amount (positive for deposit, negative for all else) */
-  chk_amount FLOAT NOT NULL,
+  chk_amount DECIMAL(14,2) NOT NULL,
   /* Date of transaction (YYYYMMDD) format */
   chk_date INT NOT NULL,
   /* Description */
